requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.4.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data processing
scipy>=1.11.0
networkx>=3.2.0

# Evaluation and metrics
scikit-surprise>=1.1.3

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Optional: for Kaggle dataset download
kaggle>=1.5.16

